Create general classes for integrators and inherit for RK. Correct tests and other instances to use the general classes
