+++ netket/experimental/driver/tdvp_schmitt.py
@@ -27,7 +27,7 @@ from netket.vqs import VariationalState, VariationalMixedState, MCState
 from netket.jax import tree_cast
 from netket.utils import timing
 
-from netket.experimental.dynamics import RKIntegratorConfig
+from netket.experimental.dynamics import IntegratorConfig
 
 
 from .tdvp_common import TDVPBaseDriver, odefun
@@ -105,7 +105,7 @@ class TDVPSchmitt(TDVPBaseDriver):
         self,
         operator: AbstractOperator,
         variational_state: VariationalState,
-        integrator: RKIntegratorConfig,
+        integrator: IntegratorConfig,
         *,
         t0: float = 0.0,
         propagation_type: str = "real",
