+++ test/dynamics/test_solvers.py
@@ -16,10 +16,18 @@ import pytest
 import numpy as np
 
 import jax
-import scipy.integrate as sci
+from functools import partial
 
-from netket.experimental.dynamics import Euler, Heun, Midpoint, RK4, RK12, RK23, RK45
-from netket.experimental.dynamics._rk_tableau_definitions import (
+from netket.experimental.dynamics import (
+    Euler,
+    Heun,
+    Midpoint,
+    RK4,
+    RK12,
+    RK23,
+    RK45,
+)
+from netket.experimental.dynamics._rk._tableau import (
     bt_feuler,
     bt_heun,
     bt_midpoint,
@@ -32,10 +40,10 @@ from netket.experimental.dynamics._rk_tableau_definitions import (
 
 from .. import common
 
-pytestmark = common.skipif_distributed
+pytestmark = common.skipif_mpi
 
 
-tableaus = {
+tableaus_rk = {
     "bt_feuler": bt_feuler,
     "bt_heun": bt_heun,
     "bt_midpoint": bt_midpoint,
@@ -47,60 +55,59 @@ tableaus = {
 }
 
 
-explicit_fixed_step_solvers = {
+fixed_step_solvers = {
     "Euler": Euler,
     "Heun": Heun,
     "Midpoint": Midpoint,
     "RK4": RK4,
 }
 
-explicit_adaptive_solvers = {
+adaptive_solvers = {
     "RK12": RK12,
     "RK23": RK23,
     "RK45": RK45,
 }
 
-tableaus_params = [pytest.param(obj, id=name) for name, obj in tableaus.items()]
-explicit_fixed_step_solvers_params = [
-    pytest.param(obj, id=name) for name, obj in explicit_fixed_step_solvers.items()
+rk_tableaus_params = [pytest.param(obj, id=name) for name, obj in tableaus_rk.items()]
+fixed_step_solvers_params = [
+    pytest.param(obj, id=name) for name, obj in fixed_step_solvers.items()
 ]
-explicit_adaptive_solvers_params = [
-    pytest.param(obj, id=name) for name, obj in explicit_adaptive_solvers.items()
+adaptive_solvers_params = [
+    pytest.param(obj, id=name) for name, obj in adaptive_solvers.items()
 ]
 
 
-@pytest.mark.parametrize("tableau", tableaus_params)
-def test_tableau(tableau: str):
+@pytest.mark.parametrize("tableau", rk_tableaus_params)
+def test_tableau_rk(tableau: str):
     assert tableau.name != ""
-    td = tableau.data
 
-    for x in td.a, td.b, td.c:
+    for x in tableau.a, tableau.b, tableau.c:
         assert np.all(np.isfinite(x))
 
-    assert td.a.ndim == 2
+    assert tableau.a.ndim == 2
     # a should be strictly upper triangular
-    np.testing.assert_array_equal(np.triu(td.a), np.zeros_like(td.a))
+    np.testing.assert_array_equal(np.triu(tableau.a), np.zeros_like(tableau.a))
     # c's should be in [0, 1]
-    assert np.all(td.c >= 0.0)
-    assert np.all(td.c <= 1.0)
+    assert np.all(tableau.c >= 0.0)
+    assert np.all(tableau.c <= 1.0)
 
-    assert len(td.order) in (1, 2)
-    assert len(td.order) == td.b.ndim
+    assert len(tableau.order) in (1, 2)
+    assert len(tableau.order) == tableau.b.ndim
 
-    assert td.a.shape[0] == td.a.shape[1]
-    assert td.a.shape[0] == td.b.shape[-1]
-    assert td.a.shape[0] == td.c.shape[0]
-    if len(td.order) == 2:
-        assert td.b.shape[0] == 2
+    assert tableau.a.shape[0] == tableau.a.shape[1]
+    assert tableau.a.shape[0] == tableau.b.shape[-1]
+    assert tableau.a.shape[0] == tableau.c.shape[0]
+    if len(tableau.order) == 2:
+        assert tableau.b.shape[0] == 2
 
 
-@pytest.mark.parametrize("method", explicit_fixed_step_solvers_params)
+@pytest.mark.parametrize("method", fixed_step_solvers_params)
 def test_fixed_adaptive_error(method):
     with pytest.raises(TypeError):
         method(dt=0.01, adaptive=True)
 
 
-@pytest.mark.parametrize("method", explicit_fixed_step_solvers_params)
+@pytest.mark.parametrize("method", fixed_step_solvers_params)
 def test_ode_solver(method):
     def ode(t, x, **_):
         return -t * x
@@ -112,8 +119,7 @@ def test_ode_solver(method):
     y0 = np.array([1.0])
     times = np.linspace(0, n_steps * dt, n_steps, endpoint=False)
 
-    sol = sci.solve_ivp(ode, (0.0, n_steps * dt), y0, t_eval=times)
-    y_ref = sol.y[0]
+    y_ref = y0 * np.exp(-(times**2) / 2)
 
     solv = solver(ode, 0.0, y0)
 
@@ -132,10 +138,7 @@ def test_ode_solver(method):
 
     # somewhat arbitrary tolerances, that may still help spot
     # errors introduced later
-    rtol = {
-        "Euler": 1e-2,
-        "RK4": 5e-4,
-    }.get(solver.tableau.name, 1e-3)
+    rtol = {"Euler": 1e-2, "RK4": 5e-4}.get(solver.tableau.name, 1e-3)
     np.testing.assert_allclose(y_t[:, 0], y_ref, rtol=rtol)
 
 
@@ -145,21 +148,20 @@ def test_ode_repr():
     def ode(t, x, **_):
         return -t * x
 
-    solver = RK23(dt=dt, adaptive=True)
-
+    solver = RK23(dt, adaptive=True)
     y0 = np.array([1.0])
     solv = solver(ode, 0.0, y0)
 
     assert isinstance(repr(solv), str)
-    assert isinstance(repr(solv._rkstate), str)
+    assert isinstance(repr(solv._state), str)
 
     @jax.jit
     def _test_jit_repr(x):
-        assert isinstance(repr(x), str)
         return 1
 
-    _test_jit_repr(solv._rkstate)
-    # _test_jit_repr(solv) # this is broken. should be fixed in the zukumft
+    _test_jit_repr(solv._state)
+    _test_jit_repr(solv.tableau)
+    # _test_jit_repr(solv)  # this is broken. should be fixed in the zukunft
 
 
 def test_solver_t0_is_integer():
@@ -169,10 +171,8 @@ def test_solver_t0_is_integer():
     def df(t, y, stage=None):
         return np.sin(t) ** 2 * y
 
-    int_config = RK23(
-        dt=0.04, adaptive=True, atol=1e-3, rtol=1e-3, dt_limits=[1e-3, 1e-1]
-    )
-    integrator = int_config(
+    init_config = RK23(dt=0.04, adaptive=True, atol=1e-3, rtol=1e-3, dt_limits=[1e-3, 1e-1])
+    integrator = init_config(
         df, 0, np.array([1.0])
     )  # <-- the second argument has to be a float
 
@@ -181,7 +181,7 @@ def test_solver_t0_is_integer():
     assert integrator.t.dtype == integrator.dt.dtype
 
 
-@pytest.mark.parametrize("solver", explicit_adaptive_solvers_params)
+@pytest.mark.parametrize("solver", adaptive_solvers_params)
 def test_adaptive_solver(solver):
     tol = 1e-7
 
@@ -195,18 +195,16 @@ def test_adaptive_solver(solver):
     y_t = []
     last_step = -1
     while solv.t <= 2.0:
-        # print(solv._rkstate)
-        if solv._rkstate.step_no != last_step:
-            last_step = solv._rkstate.step_no
+        # print(solv._state)
+        if solv._state.step_no != last_step:
+            last_step = solv._state.step_no
             t.append(solv.t)
             y_t.append(solv.y)
         solv.step()
     y_t = np.asarray(y_t)
+    t = np.asarray(t)
 
     # print(t)
-    sol = sci.solve_ivp(
-        ode, (0.0, 2.0), y0, t_eval=t, atol=0.0, rtol=tol, method="RK45"
-    )
-    y_ref = sol.y[0]
+    y_ref = y0 * np.exp(-(t**2) / 2)
 
     np.testing.assert_allclose(y_t[:, 0], y_ref, rtol=1e-5)
